pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "qemu.qmp>=0.0.5",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
osworld = [
    "numpy",
    "backoff",
    "pandas",
    "openai",
    "anthropic",
    "fastapi",
    "zhipuai",
    "groq",
    "boto3",
    "uvicorn",
    "paddleocr",
    "paddlepaddle",
    "together",
    "scikit-learn",
    "websockets",
    "tiktoken",
    "pyautogui~=0.9.54",
    "toml",
    "exa_py",
    "black",
    "certifi",
    "pytesseract",
    "google-genai",
    "python-dotenv",
    "Pillow~=11.0.0",
    "fabric",
    "gymnasium~=0.28.1",
    "requests~=2.31.0",
    "pytz~=2024.1",
    "transformers~=4.35.2",
    # changed from ~=, for newer torchvision with python 3.13 builds
    "torch>=2.5.0",
    "accelerate",
    "opencv-python~=4.8.1.78",
    "matplotlib~=3.7.4",
    "psutil~=5.9.6",
    "tqdm~=4.65.0",
    "flask~=3.0.0",
    "requests-toolbelt~=1.0.0",
    "lxml",
    "cssselect",
    "xmltodict",
    "openpyxl",
    "python-docx",
    "python-pptx",
    "pypdf",
    "PyGetWindow",
    "rapidfuzz",
    "pyacoustid",
    "pygame",
    "ImageHash",
    "scikit-image",
    "librosa",
    "pymupdf",
    "chardet",
    "playwright",
    "formulas",
    "pydrive",
    "fastdtw",
    "odfpy",
    "func-timeout",
    "beautifulsoup4",
    "dashscope",
    "google-generativeai",
    "PyYaml",
    "mutagen",
    "easyocr",
    "borb",
    "pypdf2",
    "pdfplumber",
    "wandb",
    "wrapt_timeout_decorator",
    "gdown",
    "azure-identity",
    "azure-mgmt-compute",
    "azure-mgmt-network",
    "docker",
    "loguru",
    "dotenv",
    'pyobjc; platform_system == "Darwin"',
    'pywinauto; platform_system == "Windows"',
    'pywin32; platform_system == "Windows"',
    "tldextract>=5.3.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
